+++ CHANGELOG.md
@@ -12,6 +12,9 @@ If you like this project and find it useful, please consider giving it a star on
 
 ### Added
 
+- [cli]: Added takeHeapSnapshot() and triggerGarbageCollection().
+- [LaundryWasher]: Add LaundryWasher class and Jest test.
+
 ### Changed
 
 ### Fixed